#93705: First steps in the development of the map component for the SmartCities project.
Included is a minimal extension of Leaflet's base map, simple test page and a jakefile script in which we will be adding build tasks. For now, there is only a "doc" task that uses jsdoc to create a doc folder with a javadoc-like html structure created from js files comments.